Separate order model and mapper
