+++ src/models/order.ts
@@ -1,0 +1,11 @@
+import {Side} from './index';
+
+interface Order {
+  id: string;
+  volume: number;
+  price: number;
+  timestamp: any;
+  side: Side;
+}
+
+export default Order;
